Move BaseResource to types
