+++ types/package.json
@@ -23,7 +23,8 @@
     "@tsconfig/recommended": "^1.0.3",
     "@types/uuid": "^9.0.7",
     "tslib": "^2.6.2",
-    "typescript": "^5.0.3"
+    "typescript": "^5.0.3",
+    "sequelize": "^6.31.0"
   },
   "dependencies": {
     "@notionhq/client": "^2.2.4",
