+++ types/src/shared/result.ts
